Removed the old movement scripts and renamed some classes I also moved the cell script into the model dir since it fits better there
