--- unity/Group55TestProjekt/Assets/Scripts/CombinedMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3f82caa06a076f499700df395d191df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
