+++ unity/Group55TestProjekt/Assets/Scripts/Movement.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4b92b5bf95430b14caba4f02febedcf7
+guid: a3f82caa06a076f499700df395d191df
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
