--- unity/Group55TestProjekt/Assets/Scripts/NewMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cabc3d6198ccff04cb5ac51c7d795279
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
